Gameboard and Ship complete but too messy
need to clean up a lot. setShip22() is being used while setShip() does
not work. I will just delete setShip and keep setShip22().

setShip(ship, inputArray) currently checks a couple of things:
1. is inputArray.length === ship.length ??
2. are all inputs valid coordinates on the board AND EMPTY
3. map inputArray into allIndex
4. check if allIndex is filled with either all indices in the same row
or all indices in same column
5. if its either all Horizontal or all Vertical, then set all board
indices that correspond to allIndex to point to the ship
AKA success... set the coordinate of the board to the ship

I feel like all these checks may be better placed separately as private
functions, and then called in by setShip22, but im not sure.
Also, maybe it is better for each Ship object to have this function
instead. Although this is sort of confusing bc if you do this, the ship
will give you 'possible placements' that might not be allowed by the
board currently (for ex: this spot is already taken).
So if I put this function in the ship, I will still need the Gameboard
to look through these 'possible spots the ship can occupy when being
placed', and adjust this array based on the current board....
I was thinking of it like chess pieces and this is likely what they do.
Each piece has 'possible moves' based on current possition. Then the
board/gamestate narrow down those possible moves

HOWEVER I think this might be too complicated here, so maybe just see
if it makes sense to separate all the check functions inside setShip22