In middle of a big mess while rewritting setShip
the setShip portion at the bottom that we were rewriting seems like
it forces me to write too much code.... maybe less is better even if it
is worse algorithm

HOWEVER this also makes me realize that we are doing this WAYYYYY TOO
MESSY. We are working with inputs, coordinates, and index,
and different methods expect a differnt parameter... which is impossible
to tell. I think it is best to just make all methods work with index.
if we want to, we can also have some helper methods here for outside
callers, such as inputToIndex() and isInputValid()

this is copied from comments at bottom of Gameboard.js
so.. rewrite the whole thing so it works just with index
then provide helper methods to caller to convert inputs or
coordinates into index