Wrote some simple foundation DOM functions
Now need to polish display/render functions and clean structure.
Below is a copy and paste comment i wrote at the end of my index.js.
Also rememebr all the stuff in index.js will ultimately be moved into
Game() and be inside a game loop:

TODO i think pass the player insted of the board... then getGameboard()
INSIDE the function. This way we can extract things like name inside
the function. And more IMPORTANTLY, we can add a "() => player"
to the click handler of the squares
AKA the gameboard renders will be able to tell me which player it comes
from

.... Although now that i think about it, its kinda strange that
I cant otherwise tell...
very sleepy, but i think its because we have a player -> gameboard
relationship, and the display is completely separate... so theres
no connection to the player that owns the gameboard