Finished bootleg version with working DOM &console
Need to turn all the log statements into visible 'responses'. Maybe w/
a chatbox in between the 2 player views... or something. Maybe dont turn
ALL of the console.logs tho, maybe only some.

Also figure out what to do when the game ends. We will probably throw
a pubsub event emit and catch it in index.js or something... not sure.
But we want the players to be able to choose to start a new Game.
So maybe yes do it this way, but we would have to have some Match()
object in index that runs Game(). Also we need to actually code the
endGame() inside Game properly, here is the commment i left:

For now we are not really 'ENDING' the game... we are just not rendering
or passing turn. Which means it is permanently winner's turn, but I am
not showing the results from the hits. NEED To actually end the
game somehow.... and show a screen that lets me restart the game..

have to think about this because right now all the square elements have
click listeners. But I dont want to immediately render a diff page,
bc I want ppl to be able to look at the Game after it ends.. So probably
just renderPlayerViws() but have a method that DOESNT add the
action listeners... soemthign like that